Made Changes to CLI...
With how easy StaticPHP is to use when using a launcher script, which is why it is the recommended method to use StaticPHP, I haven't really given much love to StaticPHP's CLI.

This update focuses entirely on how StaticPHP handles and parses CLI arguments.

+ Added Named Flags (--configurable_option=value)
* Original CLI Method now called Positional Arguments Method
* Cleaned up CLI parsing code to make future development easier
* Updated and Cleaned CLI Documentation on Getting Started page

Hopefully this update benefits someone, even if it is just my future self. Enjoy! :)